github: Change return type of GetWorkflowRuns to pointer
GetWorkflowRuns returned a list of WorkflowRun instances. It's expected
that these instances can be modified, and those modifications will be
reflected in any struct that embeds them. When iterating over
WorkflowRuns in workflow_runs.go, the `run` variable created by the
`range` loop is a copy of the WorkflowRun in the slice, not a reference.
This caused a bug where any modifications made to the WorkflowRun struct
would not be reflected in the outputted OpenSearch json request.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>