ci: Tolerate dispatch workflows with no echo-inputs job on non-main
This commit temporarily modifies the logic in the workflow runs command
to not fatally exit if a workflow_dispatch triggered workflow does not
have an echo-inputs job, if the head branch of the workflow is not main.
This is done to allow scraping of branches other than main while the
echo-inputs job is backported to stable branches.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>