ci: Use index verb and grab last three hours on scheduled runs
This commit updates the OpenSearch bulk requests to use the index verb
and additionally modifies the workflow scraping runs to pull the last
three hours of data from GitHub, rather than just the last hour. This
will help to alleviate issues that may arise with scheduled workflows
that fail to execute due to high workflow volume.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>