opensearch: Escape strings in document IDs
Some of the document identifiers contain string data from a field in the
document, such as the `DocumentIdentifier` field in a `FailureRate`
document. These strings are not guaranteed to be escaped for usage in a
JSON string, so this commit ensures they are formatted to not cause
trouble when attempting to index.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>